requirements.txt
alabaster==1.0.0
alembic==1.15.2
APScheduler==3.11.0
babel==2.17.0
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.8
colorama==0.4.6
dataclass-wizard==0.35.0
docutils==0.21.2
Flask==3.1.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
greenlet==3.1.1
h5py==3.13.0
idna==3.4
imagesize==1.4.1
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
llvmlite==0.44.0
lxml==5.3.2
Mako==1.3.9
MarkupSafe==3.0.2
numba==0.61.0
numpy==1.25.0
numpy-quaternion==2024.0.7
packaging==24.2
pandas==2.0.2
pillow==11.1.0
pluggy==1.5.0
psycopg2-binary==2.9.10
Pygments==2.19.1
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
pyquaternion==0.9.9
pytest==8.3.5
python-dateutil==2.8.2
python-dotenv==1.1.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
roman-numerals-py==3.1.0
scipy==1.15.2
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.6
Sphinx==8.2.3
sphinx-autodoc-typehints==3.1.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-httpdomain==1.8.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.40
tqdm==4.67.1
turtle==0.0.1
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==5.3.1
urllib3==2.0.3
Werkzeug==3.1.3
zod==0.6.1
